requirements.txt
requests
beautifulsoup4
Flask
gunicorn
google-cloud-storage==2.10.0